move payment request stuff into a subdirectory
