+++ lib/js/PKPaymentButton/index.js
