+++ lib/js/PaymentRequest/PaymentRequestUpdateEvent.js
@@ -12,8 +12,8 @@ import {
 } from './constants';
 const noop = () => {};
 
-import PaymentRequest from './PaymentRequest';
-import NativePayments from './NativePayments';
+import PaymentRequest from '.';
+import NativePayments from '../NativeBridge';
 
 // Helpers
 import {
