+++ lib/js/PaymentRequest/__mocks__/index.js
