+++ lib/js/PaymentRequest/__tests__/PaymentRequestUpdateEvent.test.js
@@ -5,8 +5,8 @@ const {
 } = require('../__mocks__');
 
 jest.mock('react-native', () => mockReactNativeIOS);
-jest.mock('../NativePayments', () => mockNativePaymentsSupportedIOS);
-const PaymentRequest = require('../PaymentRequest').default;
+jest.mock('../../NativeBridge', () => mockNativePaymentsSupportedIOS);
+const PaymentRequest = require('../').default;
 const PaymentRequestUpdateEvent = require('../PaymentRequestUpdateEvent')
   .default;
 
@@ -16,7 +16,7 @@ import {
   createInteractivePaymentRequest,
   createClosedPaymentRequest,
   createUpdatingPaymentRequest
-} from './PaymentRequest-test.js';
+} from './index.test.js';
 
 // constants
 const id = 'foo';
