+++ lib/js/PaymentRequest/__tests__/constants.test.js
