+++ lib/js/PaymentRequest/__tests__/index.test.js
@@ -5,8 +5,8 @@ const {
 } = require('../__mocks__');
 
 jest.mock('react-native', () => mockReactNativeIOS);
-jest.mock('../NativePayments', () => mockNativePaymentsSupportedIOS);
-const PaymentRequest = require('../PaymentRequest').default;
+jest.mock('../../NativeBridge', () => mockNativePaymentsSupportedIOS);
+const PaymentRequest = require('../').default;
 
 // helpers
 export function createCreatedPaymentRequest(methodData, details, options) {
