+++ lib/js/PaymentRequest/errors/__tests__/index.test.js
