+++ lib/js/PaymentRequest/helpers/__tests__/index.test.js
