+++ lib/js/PaymentRequest/helpers/index.js
