+++ lib/js/PaymentRequest/index.js
@@ -14,14 +14,14 @@ import type {
   PaymentShippingType,
   PaymentDetailsIOS,
   PaymentDetailsIOSRaw,
-} from './types';
+} from '../types';
 import type PaymentResponseType from './PaymentResponse';
 
 // Modules
 import { DeviceEventEmitter, Platform } from 'react-native';
 import uuid from 'uuid/v1';
 
-import NativePayments from './NativePayments';
+import NativePayments from '../NativeBridge';
 import PaymentResponse from './PaymentResponse';
 import PaymentRequestUpdateEvent from './PaymentRequestUpdateEvent';
