+++ lib/js/index.js
@@ -1,7 +1,2 @@
-// @flow
-
-import _PaymentRequest from './PaymentRequest';
-import { PKPaymentButton } from './PKPaymentButton';
-
-export const ApplePayButton = PKPaymentButton;
-export const PaymentRequest = _PaymentRequest;
+export { PaymentRequest } from './PaymentRequest';
+export { PKPaymentButton } from './PKPaymentButton';
